+++ web-components/src/off-webcomponents.ts
@@ -1,13 +1,9 @@
 export { setLocale, getLocale } from "./localization"
-export { OffWebcomponentsConfiguration as OffWebcomponentConfiguration } from "./components/shared/off-webcomponents-configuration"
+export { OffWebcomponentsConfiguration } from "./components/shared/off-webcomponents-configuration"
 export { RobotoffQuestion } from "./components/robotoff-question/robotoff-question"
-export { RobotoffNutrients } from "./components/robotoff-nutrients/robotoff-nutrients"
-export { RobotoffIngredientSpellcheck as RobotoffIngredients } from "./components/robotoff-ingredient-spellcheck/robotoff-ingredient-spellcheck"
+export { RobotoffNutrientExtraction } from "./components/robotoff-nutrient-extraction/robotoff-nutrient-extraction"
+export { RobotoffIngredientSpellcheck } from "./components/robotoff-ingredient-spellcheck/robotoff-ingredient-spellcheck"
 export { RobotoffContributionMessage } from "./components/robotoff-contribution-message/robotoff-contribution-message"
-
-// Will be added in the future when design is ready
-// export { KnowledgePanelsComponent } from "./components/knowledge-panels/knowledge-panels"
-
 export { DonationBanner } from "./components/donation-banner/donation-banner"
 export { MobileBadges } from "./components/mobile-badges/mobile-badges"
 export { DownloadAppQrCode } from "./components/download-app-qr-code/download-app-qr-code"
@@ -16,3 +12,6 @@ export { DeleteModal } from "./components/folksnomy/delete-modal"
 export { FolksonomyEditor } from "./components/folksnomy/folksonomy-editor"
 export { AutocompleteInput } from "./components/shared/autocomplete-input"
 export { RobotoffIngredientDetection as RobotoffCrops } from "./components/robotoff-ingredient-detection/robotoff-ingredient-detection"
+
+// Will be added in the future when design is ready
+// export { KnowledgePanelsComponent } from "./components/knowledge-panels/knowledge-panels"